+++ tests/testthat/test-sl.R
@@ -56,7 +56,7 @@ sl2_fit <- sl2$train(task)
 sl2_risk <- sl2_fit$cv_risk(loss_squared_error)
 
 test_that("Lrnr_sl can accept a pre-made stack", {
-  expect_equal(sl1_risk$mean_risk, sl2_risk$mean_risk, tolerance = 1e-2)
+  expect_equal(sl1_risk$MSE, sl2_risk$MSE, tolerance = 1e-2)
 })
 
 sl_nnls <- Lrnr_sl$new(